Cleaned up variable naming, made more consistent, added from_storage and create static methods to extents
