+++ src/silvimetric/__init__.py
@@ -1,4 +1,5 @@
 __version__ = '0.0.1'
 
-from .bounds import Bounds, Extents, create_extents
+from .bounds import Bounds
+from .extents import Extents
 from .storage import Storage
