+++ src/silvimetric/cli/__main__.py
@@ -3,7 +3,6 @@ import logging
 import json
 
 from silvimetric.app import Application
-from silvimetric.bounds import Extents
 from silvimetric.storage import Storage
 
 logger = logging.getLogger(__name__)
